test_bytestream01.cpp: Fix unused return value warning -> error gcc 11.2

+++ test/test_bytestream01.cpp
@@ -96,21 +96,21 @@ class TestByteStream01 {
         }
 
         ~TestByteStream01() {
-#ifdef USE_LIBCURL
-            int res = std::system("killall mini_httpd");
-            (void)res;
-#endif // USE_LIBCURL
+            if( jau::io::uri_tk::protocol_supported("http:") ) {
+                int res = std::system("killall mini_httpd");
+                (void)res;
+            }
         }
 
         static void httpd_start() {
-#ifdef USE_LIBCURL
-            int res = std::system("killall mini_httpd");
-            const std::string cwd = jau::fs::get_cwd();
-            const std::string cmd = "/usr/sbin/mini_httpd -p 8080 -l "+cwd+"/mini_httpd.log";
-            jau::PLAIN_PRINT(true, "%s", cmd.c_str());
-            res = std::system(cmd.c_str());
-            (void)res;
-#endif // USE_LIBCURL
+            if( jau::io::uri_tk::protocol_supported("http:") ) {
+                int res = std::system("killall mini_httpd");
+                const std::string cwd = jau::fs::get_cwd();
+                const std::string cmd = "/usr/sbin/mini_httpd -p 8080 -l "+cwd+"/mini_httpd.log";
+                jau::PLAIN_PRINT(true, "%s", cmd.c_str());
+                res = std::system(cmd.c_str());
+                (void)res;
+            }
         }
 
         static bool transfer(jau::io::ByteInStream& input, const std::string& output_fname, const size_t buffer_size=4096) {
